Expander working
Sadly, psyntax is written in itself, and so really it should require a
pre-expanded copy to work -- but since we're staging from Guile to Bicho
that's not a great option.  The meaning is that all the `syntax' and
`syntax-case' forms in the file that are not part of the (macroexpand
'(...)) bit at the end are implemented with Guile's syntax-case -- which
may mean that Guile's syntax-case is used to destructure Bicho syntax
objects.  Nasty.  We hack around it by adding an extra field to Bicho's
syntax objects to make them the same shape as Guile's (gulp).